backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.5
langchain>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0
httpx>=0.27.0
apscheduler>=3.10.4
pydantic>=2.0.0
pydantic-settings>=2.0.0
supabase>=2.4.0
